Use wider types in spin lock unit test.

+++ multicore_engine_core_tests/src/util/spin_lock_test.cpp
@@ -25,12 +25,12 @@ BOOST_AUTO_TEST_CASE(thread_safety_lock) {
 	const int thread_count = 256;
 	const int elements_per_thread = 1024;
 
-	test.reserve(thread_count * elements_per_thread);
+	test.reserve(thread_count * static_cast<size_t>(elements_per_thread));
 	for(int i = 0; i < thread_count; i++) {
 		threads.emplace_back([&]() {
 			for(int j = 0; j < elements_per_thread; ++j) {
 				std::lock_guard<mce::util::spin_lock> guard(lock);
-				test.emplace_back((static_cast<uint64_t>(i) << 32) | j);
+				test.emplace_back((static_cast<uint64_t>(i) << 32) | static_cast<uint64_t>(j));
 			}
 		});
 	}
@@ -45,13 +45,13 @@ BOOST_AUTO_TEST_CASE(thread_safety_try_lock) {
 	const int elements_per_thread = 102400;
 	std::atomic<size_t> successful_insertions{0};
 
-	test.reserve(thread_count * elements_per_thread);
+	test.reserve(thread_count * static_cast<size_t>(elements_per_thread));
 	for(int i = 0; i < thread_count; i++) {
 		threads.emplace_back([&]() {
 			for(int j = 0; j < elements_per_thread; ++j) {
 				if(lock.try_lock()) {
 					successful_insertions++;
-					test.emplace_back((static_cast<uint64_t>(i) << 32) | j);
+					test.emplace_back((static_cast<uint64_t>(i) << 32) | static_cast<uint64_t>(j));
 					lock.unlock();
 				}
 			}
